Extract PDF Merge Feature as Component
